Initialize example of time dependent elements.
It is not clear how best to create time dependent conductors in Warp.
In addition, in the MEMS main script the time-dependence is handled
through the switch wp.gridmode = 0 which is a costly operation
(see f3d.v in Warp files). Thus, this script is see how best to create
a simple time oscillating capacitor in Warp.