Plot Ez contours. Set color bar with Ezmax
The Ez contour is much more informative and helps to verify stacks are
acting properly (field free region in drift). The color bar is set to
Ezmax which is approximated using the max voltage divided by the gap
length. Ezmax is then scaled by a scale factor to be higher than this
approximted value to get any higher fields.