Update geometry and hard_edge_kappa function.
The geometry was incorrect for the lattice period. The hard edge setup
thus needed to be changed. Rather than use global variables, I decided
to use default values instead for better control. A new plot was created
to visualize the geometry on the mesh. Mesh elements added to solver
plots for better visualization.