Cost function created.
The cost function uses the solver created in the solver.py file for
solving the KV envelope equations. The solver in the cost function was
tested by added the plotting routines in the cost function and executing
the function. The outputs were identical and so the solver has been
correctly incorporated.