Add radial constraint to solver.
The radial constraint is .55mm set by the physical size of the aperture.
It would be wise to set this to a lower value, but at first pass I will
take the maximum possible excursion. The contraint is added through the
penalty function. From multiple runs, the cost is in the general range of
1-5. This justifies that relatively large 500 value.